feat: Update header navigation links
The header navigation links have been updated to include a "Contatos" link when the user is logged in and a "Sair" link for logout. When the user is not logged in, the links remain as "Entrar" and "Cadastrar". This change improves the user experience by providing relevant navigation options based on their authentication status.